Move merchandise models from model.commercial to commercial.model.merchandise.
