+++ commercial/app/controllers/Multi.scala
@@ -6,9 +6,9 @@ import commercial.model.merchandise.events.MasterclassAgent
 import commercial.model.merchandise.jobs.JobsAgent
 import commercial.model.merchandise.soulmates.SoulmatesAgent
 import commercial.model.merchandise.travel.TravelOffersAgent
+import commercial.model.merchandise.{Book, Job, LiveEvent, Masterclass, Member, MemberPair, Merchandise, TravelOffer}
 import common.{ExecutionContexts, JsonComponent}
 import model.Cached
-import model.commercial._
 import play.api.libs.json.{JsArray, Json}
 import play.api.mvc._
 import play.twirl.api.Html
