+++ commercial/app/controllers/SoulmatesController.scala
@@ -3,7 +3,7 @@ package commercial.controllers
 import common.JsonComponent
 import commercial.model.merchandise.soulmates.{SoulmatesAgent, Sample}
 import commercial.model.merchandise.soulmates.SoulmatesAgent.{menAgent, newMenAgent, newWomenAgent, womenAgent}
-import model.commercial.Member
+import commercial.model.merchandise.Member
 import model.Cached
 import play.api.mvc._
