+++ commercial/app/model/merchandise/MerchandiseAgent.scala
@@ -1,4 +1,4 @@
-package model.commercial
+package commercial.model.merchandise
 
 import commercial.model.Segment
 import common.{AkkaAgent, Logging}
