+++ commercial/app/model/merchandise/soulmates/SoulmatesFeed.scala
@@ -4,7 +4,7 @@ import java.lang.System._
 
 import commercial.model.feeds.{FeedMetaData, MissingFeedException, ParsedFeed, SwitchOffException}
 import common.{ExecutionContexts, Logging}
-import model.commercial.Member
+import commercial.model.merchandise.Member
 import play.api.libs.json.{JsValue, Json}
 
 import scala.concurrent.Future
