+++ commercial/app/model/merchandise/travel/TravelOffersApi.scala
@@ -4,7 +4,7 @@ import java.lang.System.currentTimeMillis
 
 import commercial.model.feeds.{FeedMetaData, MissingFeedException, ParsedFeed, SwitchOffException}
 import common.{ExecutionContexts, Logging}
-import model.commercial.TravelOffer
+import commercial.model.merchandise.TravelOffer
 import org.joda.time.format.DateTimeFormat
 
 import scala.concurrent.Future
