+++ commercial/app/views/books/bookCard.scala.html
@@ -1,4 +1,6 @@
-@(book: model.commercial.Book,
+@import commercial.model.merchandise.Book
+
+@(book: Book,
   clickMacro: Option[String],
   optClassNames: Option[Seq[String]] = None)
