+++ commercial/app/views/books/bookLargeCard.scala.html
@@ -1,4 +1,6 @@
-@(book: model.commercial.Book,
+@import commercial.model.merchandise.Book
+
+@(book: Book,
   clickMacro: Option[String],
   optAdvertClassNames: Option[Seq[String]] = None)
