+++ commercial/app/views/books/booksBlended.scala.html
@@ -1,4 +1,6 @@
-@(book: model.commercial.Book, clickMacro: Option[String])
+@import commercial.model.merchandise.Book
+
+@(book: Book, clickMacro: Option[String])
 
 <div class="advert-blended advert-blended--books">
     <div class="advert-blended__title">
