+++ commercial/app/views/contentapi/items.scala.html
@@ -1,4 +1,4 @@
-@import _root_.commercial.controllers.{PaidFor, SponsorType, Supported}
+@import commercial.controllers.{PaidFor, SponsorType, Supported}
 @import common.commercial.CardContent
 @import common.LinkTo
 @import fragments.commercial.paidForMeta
