+++ commercial/app/views/jobs/jobCard.scala.html
@@ -1,4 +1,6 @@
-@(job: model.commercial.Job,
+@import commercial.model.merchandise.Job
+
+@(job: Job,
   clickMacro: Option[String],
   optClassNames: Option[Seq[String]] = None)
