+++ commercial/app/views/jobs/jobLargeCard.scala.html
@@ -1,4 +1,6 @@
-@(job: model.commercial.Job,
+@import commercial.model.merchandise.Job
+
+@(job: Job,
   clickMacro: Option[String],
   optAdvertClassNames: Option[Seq[String]] = None)
