+++ commercial/app/views/jobs/jobs.scala.html
@@ -1,11 +1,12 @@
+@import commercial.model.merchandise.Job
 @import commercial.model.merchandise.jobs.JobSector
-@(jobs: Seq[model.commercial.Job],
+@import views.html.fragments.commercial.containerWrapper
+@import views.html.jobs.jobCard
+
+@(jobs: Seq[Job],
   jobSectors: Seq[JobSector],
   omnitureId: Option[String],
   clickMacro: Option[String])(implicit request: RequestHeader)
-
-@import views.html.fragments.commercial.containerWrapper
-@import views.html.jobs.jobCard
 
 @containerWrapper(Seq("legacy", "jobs", "tone-jobs"),
                   omnitureId.map(id => s"merchandising | jobs | ${id}"),
