+++ commercial/app/views/jobs/jobsBlended.scala.html
@@ -1,4 +1,6 @@
-@(job: model.commercial.Job, clickMacro: Option[String])
+@import commercial.model.merchandise.Job
+
+@(job: Job, clickMacro: Option[String])
 
 <div class="advert-blended advert-blended--jobs">
     <div class="advert-blended__title">
