+++ commercial/app/views/liveevents/liveEvent.scala.html
@@ -1,5 +1,7 @@
-@(event: model.commercial.LiveEvent, omnitureId: Option[String], clickMacro: Option[String])(implicit request: RequestHeader)
+@import commercial.model.merchandise.LiveEvent
 @import views.support.{ImgSrc, Item300}
+
+@(event: LiveEvent, omnitureId: Option[String], clickMacro: Option[String])(implicit request: RequestHeader)
 
     <aside class="adverts adverts--manual adverts--legacy-inline adverts--tone-membership adverts--membership" data-link-name="merchandising | event | @omnitureId">
         <header class="adverts__header">
