+++ commercial/app/views/masterclasses/masterclassCard.scala.html
@@ -1,4 +1,4 @@
-@import model.commercial.Masterclass
+@import commercial.model.merchandise.Masterclass
 @import views.support.Item300
 
 @(event: Masterclass,
