+++ commercial/app/views/masterclasses/masterclassesBlended.scala.html
@@ -1,4 +1,6 @@
-@(event: model.commercial.Masterclass, clickMacro: Option[String])
+@import commercial.model.merchandise.Masterclass
+
+@(event: Masterclass, clickMacro: Option[String])
 
 @import conf.Configuration
