+++ commercial/app/views/soulmates/soulmateCard.scala.html
@@ -1,4 +1,6 @@
-@(soulmate: model.commercial.Member,
+@import commercial.model.merchandise.Member
+
+@(soulmate: Member,
   clickMacro: Option[String],
   optClassNames: Option[Seq[String]] = None)
