+++ commercial/app/views/soulmates/soulmatesBlended.scala.html
@@ -1,4 +1,6 @@
-@(members: Seq[model.commercial.Member],
+@import commercial.model.merchandise.Member
+
+@(members: Seq[Member],
   clickMacro: Option[String])
 
 @import conf.Configuration
