+++ commercial/app/views/travel/travelBlended.scala.html
@@ -1,6 +1,6 @@
-@(offer: model.commercial.TravelOffer, clickMacro: Option[String])
+@import commercial.model.merchandise.TravelOffer
 
-@import conf.Configuration
+@(offer: TravelOffer, clickMacro: Option[String])
 
 <div class="advert-blended">
     <div class="advert-blended__title">
