+++ commercial/app/views/travel/travelCard.scala.html
@@ -1,4 +1,6 @@
-@(offer: model.commercial.TravelOffer,
+@import commercial.model.merchandise.TravelOffer
+
+@(offer: TravelOffer,
   clickMacro: Option[String],
   optClassNames: Option[Seq[String]] = None)
