+++ commercial/app/views/travel/travelLargeCard.scala.html
@@ -1,4 +1,6 @@
-@(offer: model.commercial.TravelOffer,
+@import commercial.model.merchandise.TravelOffer
+
+@(offer: TravelOffer,
   clickMacro: Option[String],
   optAdvertClassNames: Option[Seq[String]] = None)
