+++ commercial/test/model/books/MagentoBestsellersFeedTest.scala
@@ -1,6 +1,6 @@
 package commercial.model.merchandise.books
 
-import model.commercial.Book
+import commercial.model.merchandise.Book
 import org.scalatest.{DoNotDiscover, FlatSpec, Matchers}
 import test.ConfiguredTestSuite
