+++ commercial/test/model/events/SingleEventbriteMasterclassParsingTest.scala
@@ -1,6 +1,6 @@
 package commercial.model.merchandise.events
 
-import model.commercial.Masterclass
+import commercial.model.merchandise.Masterclass
 import commercial.model.merchandise.events.Eventbrite.{Event, Venue}
 import org.scalatest.{FlatSpec, Matchers, OptionValues}
 import play.api.libs.json._
