+++ commercial/test/model/soulmates/Fixtures.scala
@@ -1,6 +1,6 @@
 package commercial.model.merchandise.soulmates
 
-import model.commercial.{Member, Man, Woman}
+import commercial.model.merchandise.{Member, Man, Woman}
 
 object Fixtures {
